Start rework to load book image one at a time.
`get_images` is still the main API call, but we now have a book image
context to keep the Archive open and to allow us to query with the
context for one book at a time.

After this, it will be time to integrate the context with the UI app and
then drop the full image load API.

Of note: The get book cover function is untouched. This is because the
flow is slightly different for that and we can apply some optimizations
for that.